third_party/WebKit/LayoutTests/platform/win/external/wpt/webmessaging/Transferred_objects_unusable.sub-expected.txt

This is a testharness.js-based test.
PASS Test Description: Objects listed in transfer are transferred, not just cloned, meaning that they are no longer usable on the sending side. 
PASS MessageChannel is supported. 
Harness: the test ran to completion.
